Stop the export crashing on a position with no paid cards
The first real gallery broke the export outright: 187 cards pulled, and
then every table written empty. A position whose cards all arrived free
- eight of them were rewards - divides cost by zero paid cards, which
gives pandas NA, and .astype(float) raises on NA rather than passing it
through. One player with only reward cards was enough to lose the whole
run.

Every division in the transform layer had the same latent fault, so all
fifteen now go through one numeric() helper that coerces instead of
raising, and a test builds a position of only free cards.

Two changes to the price tape, which returned nothing for all 154
positions:

- It now says why. "Too many failures" without the reason is useless in
  a log, so the first rejection is logged verbatim and every module's
  first problem appears in the end-of-run summary.
- It degrades instead of giving up. The full query asks for the sale
  venue, a non-null union: if the server cannot resolve it for one sale,
  the whole request fails and that player gets no tape. A rejection now
  drops to a query asking only for dates and prices. Those sales are
  marked UNKNOWN and count only towards the broad fair-value sets, never
  as confirmed secondary-market prints.

Also skips common and custom-series rarities, which have no tape worth
polling and were spending the run's call budget.

+++ sorare-project/sorare_portfolio/transform/essence.py
@@ -10,6 +10,8 @@ from __future__ import annotations
 
 import pandas as pd
 
+from .numbers import numeric
+
 
 def essence_tables(
     connection, positions: pd.DataFrame, settings: dict | None = None
@@ -96,11 +98,11 @@ def essence_tables(
         .reset_index()
     )
     if not by_draw.empty:
-        by_draw["eur_per_1000"] = (
+        by_draw["eur_per_1000"] = numeric(
             by_draw["card_value_eur"] / by_draw["essence_spent"].replace(0, pd.NA) * 1000
-        ).astype(float).round(2)
+        ).round(2)
         for column in ("card_value_eur", "avg_value_eur", "median_value_eur", "essence_spent"):
-            by_draw[column] = by_draw[column].astype(float).round(2)
+            by_draw[column] = numeric(by_draw[column]).round(2)
 
     return {
         "essence_ledger": ledger.drop(columns=["value_per_card_eur"], errors="ignore"),
