+++ sorare-project/sorare_portfolio/transform/investments.py
@@ -9,6 +9,8 @@ from __future__ import annotations
 
 import pandas as pd
 
+from .numbers import numeric
+
 SCENARIOS = (("bear", "bear"), ("base", "base"), ("bull", "bull"), ("extreme", "extreme"))
 
 
@@ -47,16 +49,16 @@ def investment_table(connection, positions: pd.DataFrame) -> pd.DataFrame:
 
     # Fall back to your own entry when the market has nothing to say yet.
     reference = investments["current_price_eur"].fillna(investments["avg_entry"])
-    investments["reference_price_eur"] = reference.astype(float).round(2)
-    investments["expected_upside_pct"] = (
+    investments["reference_price_eur"] = numeric(reference).round(2)
+    investments["expected_upside_pct"] = numeric(
         (investments["expected_price_eur"] / reference.replace(0, pd.NA) - 1) * 100
-    ).astype(float).round(1)
+    ).round(1)
     quantity = investments["quantity"].fillna(investments["held_cards"]).fillna(0)
-    investments["expected_profit_eur"] = (
+    investments["expected_profit_eur"] = numeric(
         (investments["expected_price_eur"] - reference.fillna(0)) * quantity
-    ).astype(float).round(2)
+    ).round(2)
 
-    upside = (investments["bull_price"].fillna(investments["target_price"]) - reference).astype(float)
-    downside = (reference - investments["downside_price"].fillna(0)).astype(float)
+    upside = numeric(investments["bull_price"].fillna(investments["target_price"]) - reference)
+    downside = numeric(reference - investments["downside_price"].fillna(0))
     investments["risk_reward_ratio"] = (upside / downside.replace(0, pd.NA)).round(2)
     return investments
